pyproject.toml
[project]
name = "pyMOA"
version = "0.1.0"
description = "A simple library for searching the web and using internet search engines."
authors = [
  { name="mehr32", email="qxc9966@protonmail.com" }
]
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.7"

dependencies = [
    "lxml",
    "requests",
    "python-dateutil"
]
license-files = ["LICEN[CS]E*"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
  "Topic :: Internet :: WWW/HTTP :: Search",
  "Topic :: Software Development :: Libraries :: Python Modules"
]


[build-system]

requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]


#[project.urls]
# Homepage = "https://github.com/pypa/sampleproject"
# Issues = "https://github.com/pypa/sampleproject/issues"
